Revert "systeminfo plugin: cancel use longlong"
This reverts commit a699b7f2bd83c7873b88048bcf721b233d3b2197.

+++ plugins/systeminfo/systeminfothread.h
@@ -18,7 +18,7 @@ public:
     explicit SysteminfoThread(QObject *parent = 0);
 
 signals:
-    void networkSpeedChanged(const int tx, const int rx);
+    void networkSpeedChanged(const quint64 tx, const quint64 rx);
 
 protected:
     void run() Q_DECL_OVERRIDE;
