+++ plugins/systeminfo/systeminfowidget.cpp
@@ -39,7 +39,7 @@ SystemInfoWidget::SystemInfoWidget(QWidget *parent) : QWidget(parent)
 
     SysteminfoThread *systeminfoThread = new SysteminfoThread;
 
-    connect(systeminfoThread, &SysteminfoThread::networkSpeedChanged, this, [=] (const int tx, const int rx) {
+    connect(systeminfoThread, &SysteminfoThread::networkSpeedChanged, this, [=] (const quint64 tx, const quint64 rx) {
         m_tx->setText(converSpeed(tx));
         m_rx->setText(converSpeed(rx));
     });
